test: add unit tests for document processing - Setup Jest mocks for external dependencies
