+++ src/document.ts
@@ -11,4 +11,3 @@ async function main() {
 main().catch((error) => {
   console.error("Error:", error);
 });
-
